Updating double values to long values in bpmn samples

+++ modules/samples/product/src/main/resources/bpmn/PerformCalculationBPMN/NumberAdder.java
@@ -3,16 +3,17 @@ package performCalculationBPMN;
 import org.activiti.engine.delegate.DelegateExecution;
 import org.activiti.engine.delegate.JavaDelegate;
 
+import java.lang.Long;
+
 
 public class NumberAdder implements JavaDelegate {
 
     @Override
     public void execute(DelegateExecution execution) throws Exception {
 
-        Double number1 = (Double) execution.getVariable("number1");
-        Double number2 = (Double) execution.getVariable("number2");
-
-        Double result = number1 + number2;
+        Long number1 = (Long) execution.getVariable("number1");
+        Long number2 = (Long) execution.getVariable("number2");
+        Long result = number1 + number2;
 
         execution.setVariable("result", result);
